+++ apps/webapp/app/routes/resources.orgs.$organizationSlug.projects.v3.$projectParam.runs.$runParam.spans.$spanParam/route.tsx
@@ -593,6 +593,14 @@ function RunBody({
                       : "–"}
                   </Property.Value>
                 </Property.Item>
+                <Property.Item>
+                  <Property.Label>Run ID</Property.Label>
+                  <Property.Value>{run.friendlyId}</Property.Value>
+                </Property.Item>
+                <Property.Item>
+                  <Property.Label>Internal ID</Property.Label>
+                  <Property.Value>{run.id}</Property.Value>
+                </Property.Item>
               </Property.Table>
             </div>
           ) : tab === "context" ? (